Add contour on game loop and question loop
The whole world is being passed around the FSM accumulator. It doesn't
look elegant however the same event streams will always yield the same
sequence of world states so we can see the FSM as a pure function taking
in a sequence of events from the world and being side efects free.
Side effects will have to be hidden in the stage actors which very non
pure.

The world data structure is a bit complex as evidenced by the
run-question let. The questions list referencing is a bit convoluted with a master
question list which is referenced by a round question lineup which is
referenced byt the round state. That should probably be better managed.