Add a webserver for communicating with the outside
Quizmaster is a console with input and output. Main display and player
lights are outputs. Outputs will have to be modeled as websockets or
SSE.
Add a webserver for communicating with the outside

Quizmaster is a console with input and output. Main display and player
lights are outputs. Outputs will have to be modeled as websockets or
SSE.