Add stage lights and displays broadcasts
Turns out we don't need player lights and a main screen broadcasts, we
can do simplify with just one endpoint where we route all relevant
events.

Quizmaster is special and gets its own messages.

The organization of the FSM also changes somewhat so that event fns are
pure-ish accumulator functions and the state fns are reporters.
Stage exposes a channel for the reporters to send events. Stage will
then distribute to the listeners.

Could greatly benefit of specs,comment annotations are crappy.